requirements.txt
aiogram==3.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
alembic==1.13.1
pytz==2024.1
python-dateutil==2.8.2
